Add functionality to save/load games
In this new version, players can now choose to save their game at any point in the game. When the program is first started, the program prompts for your name and checks to see if you already have a saved game, and asks if you'd like to load a saved game or start a new one. Finally, if a player already has 1 saved game and wants to save again, the program gives the option of either overwriting a saved game, or creating a new saved game.